☀️ 20% – A u8 parameter truncated 0x301 to 1 and faked a seven-halfword near miss
resource_38b:0240 declared Func_020012e8 as taking u8, so `Func_020012e8(0x301)`
compiled to `movs r0, #1`. The reference loads a pool word into r0 at that call,
which is what a constant too wide for an immediate looks like, so the argument
is the whole 0x301 and the parameter is not a u8.

The row scored 7 differing halfwords with the truncation and scores 36 without
it. The 7 was the artefact: it read as a near-miss and would have sent the next
reader at a row that is nowhere near. Widened, the residual is honest and is the
constant-rematerialisation family -- the reference reloads the pool word at both
uses where we hold it in r5 and copy.

Same call venus made on resource_375: a correctness fix stands when the bytes
prove the reading, even where the score gets worse.

+++ semantic/resource_38b_c_02000240.c
@@ -4,7 +4,11 @@ extern unsigned char *Data_03001ebc;
 
 extern void Func_020012fc(void);
 extern void Func_0200137a(s32);
-extern s32 Func_020012e8(u8);
+/* 引数は u8 ではない。参照は 0x301 をプール語から渡す。
+ * NOT u8. The reference loads a pool word into r0 before this call, so the
+ * argument is the whole 0x301; declared u8 it truncated to 1 and the row
+ * scored 7 differing halfwords, which read as a near-miss it is not. */
+extern s32 Func_020012e8(s32);
 extern void Func_020013a4(s32, s32);
 extern void Func_02001312(s32);
 extern void Func_02001336(void);
